Added fast sys restarts, run_server, debug.
A constant is now associated with threading (`syspq`) that determines
how many times a thread can restart within its scheduled quantum. Higher
values will benefit threads that make large amounts of non-blocking
syscalls, but carry the risk of starvation (for the same reason as an
infinite loop in a thread).

Also added is the `debug` field, which, if set, turns on one debug print
in the source (so far). That print detects and prints thread death,
which is extremely handy for seeing errors (since the corouting
mechanism normally squelches them).

libthread.run_server is a function which accepts a SyncChannel and
either a function or table (mapping to functions). It is explicitly
there to ease the development of server main loops.