Added thr_grant, improved cap translation.
The thr_grant syscall exists to inject a cap into another thread (to
which the capability is held). If that thread is the current thread,
this is a no-op--the caller already expressed access to the cap by
providing it to the syscall. For other threads, however, the change is
inobservable; the return from the syscall is a singlet which identifies
the cap in the *grantee's* cap space--it is unusable by the granter. The
granter must somehow arrange to send the singlet value to the grantee.
This could be done using the normal (blocking) SyncChannels (which
automatically grant passed caps anyway), but the real intent of this
mechanism is to ease architecture of solutions which allow value passing
between threads using well-designed side channels. (See
test_run_server.lua in this commit, which passes the return SyncChannel
for a run_server call to another thread that returns "later" by granting
the cap and passing the singlet in a scoped variable.)

To make this work, all singlet values are passed through the
supervisor/user boundary unmodified. Singlet values should never need to
be granted, anyway--they are unique so long as they are reachable, and
Lua's memory guarantees hold.